% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{clear_collator_cache}
\alias{clear_collator_cache}
\alias{collator_cache}
\alias{get_collator_cache_dir}
\title{Manage the collator Cache}
\usage{
clear_collator_cache(force = FALSE)

get_collator_cache_dir()
}
\arguments{
\item{force}{Logical. If `FALSE` (the default), an interactive prompt will ask
for confirmation before deleting files.}
}
\value{
Invisibly returns `NULL`. Used for side effects.

A character string with the path to the cache directory.
}
\description{
Helper functions to interact with the directory used by `collator` to cache
cloned GitHub repositories.
}
\section{Functions}{
\itemize{
\item \code{clear_collator_cache()}: Clear all repositories from the cache directory.

\item \code{get_collator_cache_dir()}: Get the path to the cache directory.

}}
\examples{
\dontrun{
# See where collator stores cloned repositories
get_collator_cache_dir()
}
}
